requirements.txt
# 数据获取
akshare>=1.12.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# 技术指标
ta-lib>=0.4.28
pandas-ta>=0.3.14b

# 深度学习
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# 机器学习工具
scikit-learn>=1.3.0

# 数据库
sqlalchemy>=2.0.0

# API框架
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# 任务调度
apscheduler>=3.10.0

# 配置管理
pyyaml>=6.0
python-dotenv>=1.0.0

# 可视化
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# 工具库
tqdm>=4.66.0
requests>=2.31.0

# 日志
loguru>=0.7.0

# 测试
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# 代码质量
black>=23.0.0
flake8>=6.1.0
isort>=5.12.0

# Jupyter (可选)
jupyter>=1.0.0
ipykernel>=6.25.0
